+++ dealloc_hook.m
@@ -1,10 +1,11 @@
+#import <Cocoa/Cocoa.h>
 #import "dealloc_hook.h"
 #include "_cgo_export.h"
 #import <objc/runtime.h>
 
 
 @interface Parasite : NSObject
-@property(nonatomic, assign) int hookId;
+@property(nonatomic, assign) long hookId;
 @end
 
 @implementation Parasite
@@ -16,8 +17,7 @@
 
 static void *kDeallocHookAssociation = &kDeallocHookAssociation;
 
-void Dealloc_AddHook(void* ptr, int hookId)
-{
+void Dealloc_AddHook(void* ptr, long hookId) {
     Parasite *parasite = [Parasite alloc];
     parasite.hookId = hookId;
     objc_setAssociatedObject((NSObject*)ptr, &kDeallocHookAssociation, parasite, OBJC_ASSOCIATION_RETAIN_NONATOMIC);
